Rename some file names
